requirements.txt
asgiref==3.9.1
certifi==2025.7.9
channels==4.2.2
Django==5.2.4
django-cors-headers==4.7.0
django-filter==25.1
django-jazzmin==3.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-extra-fields==3.7.0
drf-yasg==1.21.10
environs==14.2.0
filetype==1.2.0
inflection==0.5.1
marshmallow==4.0.0
packaging==25.0
pillow==11.3.0
PyJWT==2.9.0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.2.0
